#### Summary
Interface "Crafting GUI filter: by bodypart coverage; Add filter by layer (for crafting GUI, AIM, inventory etc.);;; make them two lines in the changelog, please"

#### Purpose of change

![image](https://github.com/user-attachments/assets/c63b4ccf-3209-4a25-9fee-7244baa34d54)
_Crafting GUI filter._

 - Fixes #63091 (This is the last part! I am proud of myself 😏.)

#### Describe the solution

`coverage` in Crafting GUI

 - Copy and paste (mostly) `covers` from `item_search.cpp` (used by e.g. AIM) to crafting GUI.
https://github.com/CleverRaven/Cataclysm-DDA/blob/c2fdece67993d2cea49d09efbb1684829230cdc7/src/item_search.cpp#L90-L91
 - Use the prefix `v:` just as in AIM.
 - Started a pattern of allowing each search key to define "one time job" to prepare the search. In this case, prepare bodypart ids from the user's filter string. Then, in the main loop, use that prepared data.

`layer`
 - Make a filter similar to `coverage` filter.
 - Use prefix `e:` in both, crafting GUI and AIM.

#### Describe alternatives you've considered

Move code from `item_search.cpp` into a common function and use it at both places.

#### Testing

Note: `d:` is the slow full-text search (see #63091) and `v:` is the new filter filtering by bodypart coverage.

refactor

 - Tested the refactor by searching `v:head` in AIM

coverage

 - Search `v:head`, `v:fingers`, `v:heel` etc. in the crafting GUI.
 - Searching for head armour that is close to skin with `v:head, d:close to` now takes 3 to 4 seconds. Took over 50 seconds before with `d:head, d:close to`.
 - Results are more accurate (and so much faster) if you search for a torso covering armour with `v:torso` than with `d:torso`.
   - Example found by `d:torso` and correctly not found by `v:torso` is `tempered steel chainmail sheet`. It can be used to craft `exoskeleton torso armor` which is in the full description.
 - `v:head, v:torso` returns turtleneck and such.

layer

 - Search `e: close to`, `e: normal`, `e: outer`, `e: strapp` in crafting GUI, AIM, inventory. 
 - Searching for head armour that is close to skin with `v:head, e:close to` is instant and returns the same as `v:head, d:close to` which takes 3 to 4 seconds.

#### Additional context