@Brambor 2024-12-05
I can improve it a tiny bit.