@Brambor 2024-12-05
Update: I made the alternative approach, updated **Describe the solution**