@Brambor 2024-12-05
Added filtering by layer, updated OP.